Create feature folder "resource-access", extract ResourceAccessOverview
